Moving the decorator, creating a custom exception, adapting decorator.
